refactor: move position and range modules
