+++ packages_rs/nextclade/src/analyze/aa_changes.rs
@@ -3,6 +3,8 @@ use crate::analyze::aa_sub::AaSub;
 use crate::analyze::nuc_del::NucDelRange;
 use crate::analyze::nuc_sub::NucSub;
 use crate::coord::coord_map_cds_to_global::cds_codon_pos_to_ref_range;
+use crate::coord::position::{AaRefPosition, NucRefGlobalPosition, PositionLike};
+use crate::coord::range::{have_intersection, AaRefRange, NucRefGlobalRange};
 use crate::gene::cds::Cds;
 use crate::gene::gene::GeneStrand;
 use crate::io::aa::Aa;
@@ -13,8 +15,6 @@ use crate::io::nuc::Nuc;
 use crate::translate::complement::reverse_complement_in_place;
 use crate::translate::translate_genes::{CdsTranslation, Translation};
 use crate::utils::collections::extend_map_of_vecs;
-use crate::utils::position::{AaRefPosition, NucRefGlobalPosition, PositionLike};
-use crate::utils::range::{have_intersection, AaRefRange, NucRefGlobalRange};
 use either::Either;
 use eyre::Report;
 use itertools::{Itertools, MinMaxResult};
