+++ packages_rs/nextclade/src/analyze/aa_del.rs
@@ -1,6 +1,6 @@
 use crate::analyze::aa_sub::AaSub;
+use crate::coord::position::AaRefPosition;
 use crate::io::aa::{from_aa, Aa};
-use crate::utils::position::AaRefPosition;
 use serde::{Deserialize, Serialize};
 use std::fmt::{Display, Formatter};
