+++ packages_rs/nextclade/src/analyze/find_aa_motifs.rs
@@ -1,9 +1,9 @@
 use crate::analyze::find_aa_motifs_changes::AaMotifsMap;
 use crate::analyze::virus_properties::{AaMotifsDesc, CountAaMotifsGeneDesc};
+use crate::coord::position::AaRefPosition;
+use crate::coord::range::{intersect_or_none, AaRefRange};
 use crate::io::aa::from_aa_seq;
 use crate::translate::translate_genes::{CdsTranslation, Translation};
-use crate::utils::position::AaRefPosition;
-use crate::utils::range::{intersect_or_none, AaRefRange};
 use eyre::{eyre, Report, WrapErr};
 use itertools::Itertools;
 use regex::Regex;
