+++ packages_rs/nextclade/src/analyze/find_private_aa_mutations.rs
@@ -2,6 +2,7 @@ use crate::analyze::aa_del::AaDel;
 use crate::analyze::aa_sub::AaSub;
 use crate::analyze::is_sequenced::is_aa_sequenced;
 use crate::analyze::letter_ranges::GeneAaRange;
+use crate::coord::position::{AaRefPosition, PositionLike};
 use crate::gene::cds::Cds;
 use crate::io::aa::Aa;
 use crate::io::gene_map::GeneMap;
@@ -9,7 +10,6 @@ use crate::io::letter::Letter;
 use crate::translate::translate_genes::Translation;
 use crate::tree::tree::AuspiceTreeNode;
 use crate::utils::collections::concat_to_vec;
-use crate::utils::position::{AaRefPosition, PositionLike};
 use itertools::Itertools;
 use serde::{Deserialize, Serialize};
 use std::collections::{BTreeMap, BTreeSet};
