+++ packages_rs/nextclade/src/coord/mod.rs
@@ -2,3 +2,5 @@ pub mod coord_map;
 pub mod coord_map_cds_to_global;
 pub mod coord_map_global;
 pub mod coord_map_local;
+pub mod position;
+pub mod range;
