+++ packages_rs/nextclade/src/gene/genotype.rs
@@ -1,7 +1,7 @@
+use crate::coord::position::NucRefGlobalPosition;
 use crate::io::letter::Letter;
 use crate::io::parse_pos::parse_pos;
 use crate::make_error;
-use crate::utils::position::NucRefGlobalPosition;
 use eyre::{Report, WrapErr};
 use lazy_static::lazy_static;
 use regex::Regex;
