+++ packages_rs/nextclade/src/qc/qc_rule_private_mutations.rs
@@ -1,11 +1,11 @@
 use crate::analyze::find_private_nuc_mutations::PrivateNucMutations;
 use crate::analyze::letter_ranges::NucRange;
 use crate::analyze::nuc_del::NucDel;
+use crate::coord::position::PositionLike;
+use crate::coord::range::Range;
 use crate::io::nuc::Nuc;
 use crate::qc::qc_config::QcRulesConfigPrivateMutations;
 use crate::qc::qc_run::{QcRule, QcStatus};
-use crate::utils::position::PositionLike;
-use crate::utils::range::Range;
 use num::traits::clamp_min;
 use serde::{Deserialize, Serialize};
