+++ packages_rs/nextclade/src/run/nextclade_run_one.rs
@@ -14,6 +14,7 @@ use crate::analyze::pcr_primer_changes::get_pcr_primer_changes;
 use crate::analyze::pcr_primers::PcrPrimer;
 use crate::analyze::phenotype::calculate_phenotype;
 use crate::analyze::virus_properties::{PhenotypeData, VirusProperties};
+use crate::coord::range::AaRefRange;
 use crate::io::aa::Aa;
 use crate::io::gene_map::GeneMap;
 use crate::io::letter::Letter;
@@ -27,7 +28,6 @@ use crate::translate::translate_genes::Translation;
 use crate::tree::tree::AuspiceTree;
 use crate::tree::tree_find_nearest_node::tree_find_nearest_nodes;
 use crate::types::outputs::{NextalignOutputs, NextcladeOutputs, PhenotypeValue};
-use crate::utils::range::AaRefRange;
 use eyre::Report;
 use itertools::Itertools;
 use std::collections::BTreeMap;
