+++ packages_rs/nextclade/src/translate/frame_shifts_translate.rs
@@ -1,10 +1,10 @@
 use crate::coord::coord_map_cds_to_global::cds_range_to_ref_ranges;
 use crate::coord::coord_map_local::CoordMapLocal;
+use crate::coord::position::{NucAlnLocalPosition, PositionLike};
+use crate::coord::range::{AaRefRange, NucAlnLocalRange, NucRefGlobalRange, Range};
 use crate::gene::cds::Cds;
 use crate::io::letter::Letter;
 use crate::io::nuc::Nuc;
-use crate::utils::position::{NucAlnLocalPosition, PositionLike};
-use crate::utils::range::{AaRefRange, NucAlnLocalRange, NucRefGlobalRange, Range};
 use eyre::Report;
 use itertools::Itertools;
 use serde::{Deserialize, Serialize};
