+++ packages_rs/nextclade/src/translate/translate_genes.rs
@@ -5,6 +5,8 @@ use crate::align::remove_gaps::remove_gaps_in_place;
 use crate::analyze::count_gaps::GapCounts;
 use crate::coord::coord_map_global::CoordMapGlobal;
 use crate::coord::coord_map_local::CoordMapLocal;
+use crate::coord::position::PositionLike;
+use crate::coord::range::{AaRefRange, Range};
 use crate::gene::cds::Cds;
 use crate::gene::gene::Gene;
 use crate::io::aa::Aa;
@@ -17,8 +19,6 @@ use crate::translate::translate::translate;
 use crate::types::outputs::PeptideWarning;
 use crate::utils::collections::{first, last};
 use crate::utils::error::report_to_string;
-use crate::utils::position::PositionLike;
-use crate::utils::range::{AaRefRange, Range};
 use crate::{make_error, make_internal_report};
 use eyre::Report;
 use indexmap::IndexMap;
