+++ packages_rs/nextclade/src/tree/tree.rs
@@ -1,8 +1,8 @@
+use crate::coord::position::{AaRefPosition, NucRefGlobalPosition};
 use crate::io::aa::Aa;
 use crate::io::fs::read_file_to_string;
 use crate::io::json::json_parse;
 use crate::io::nuc::Nuc;
-use crate::utils::position::{AaRefPosition, NucRefGlobalPosition};
 use eyre::{Report, WrapErr};
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
