+++ packages_rs/nextclade/src/types/outputs.rs
@@ -11,12 +11,12 @@ use crate::analyze::nuc_del::NucDelRange;
 use crate::analyze::nuc_sub::NucSub;
 use crate::analyze::pcr_primer_changes::PcrPrimerChange;
 use crate::coord::coord_map_global::CoordMapGlobal;
+use crate::coord::range::{AaRefRange, NucRefGlobalRange};
 use crate::io::json::json_parse;
 use crate::io::nuc::Nuc;
 use crate::qc::qc_run::QcResult;
 use crate::translate::frame_shifts_translate::FrameShift;
 use crate::translate::translate_genes::Translation;
-use crate::utils::range::{AaRefRange, NucRefGlobalRange};
 use eyre::Report;
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
